Add support for disabling list items
When using this component as a multiple selection dropdown, disabling
items is important. For example, if you were populating a list of
countries you'd visited, it wouldn't make sense to select "Italy"
twice. It should naturally be disabled.

* To support this I had to move to the list to be `<button>` based instead
of `<a>` based because `<a>`s do not support the `disabled` attribute.
After looking at it, I think the `<button>` is the more accurate option
anyways because we were never accepting anything as an `href` and always
forcing it to be `#`.

Questions

* I'm still debating the API for this. Currently I"m going with a semantic
type API by calling it `selectedValues` and then just acting by disabling
those elements. But, I wonder if it might be better to call it something
more explicit like `itemsToDisable`.

* I'm also worried that keyboard handing feels a little clunky. It should
just skip the item, but a disabled item is selectable with the arrow keys.